#1013 #1019 CI red recovery: PG test naming + raw-db budget + pipeline_stages readonly + review_automation INT4 decode
- Rename 3 PG-requiring tests to include _pg substring so the CI filter
  (cargo test --skip _pg --skip pg_ --skip postgres) excludes them from
  the non-PG job:
  * policy_tick_advisory_lock_blocks_split_brain_across_pools
    -> policy_tick_pg_advisory_lock_blocks_split_brain_across_pools
  * github_sync_advisory_lock_blocks_split_brain_across_pools
    -> github_sync_pg_advisory_lock_blocks_split_brain_across_pools
  * agent_quality_rollup_job_runs_one_tick
    -> agent_quality_rollup_pg_job_runs_one_tick
  Also update the reference in docs/multimachine-failover.md.
- Bump RAW_DB_BUDGET 189 -> 193 in
  engine::ops::tests::policies_raw_db_count_stays_within_budget to cover
  the 4 new callsites added by the R4 campaign. Migration to typed
  facades is a follow-up.
- Rework pipeline_stages_pg_only_without_sqlite_mirror to match #1097's
  file-canonical readonly guard: PUT and DELETE now assert HTTP 405 with
  the file-canonical error body; GET still returns 200 with an empty
  list for the test repo; sqlite mirror stays empty.
- Change the SELECT 1 probe in
  review_automation_pg_repeated_test_database_lifecycle_releases_admin_pool
  from i64 to i32 — integer literals in PG come back as INT4, so
  decoding into i64 produced a sqlx type-mismatch error.

+++ src/engine/ops/review_automation_ops.rs
@@ -1551,7 +1551,10 @@ mod tests {
             let test_db = TestDatabase::create().await;
             let pool = test_db.migrate().await;
 
-            let one: i64 = sqlx::query_scalar("SELECT 1")
+            // #1019: `SELECT 1` returns INT4 in postgres by default, so
+            // decode as i32. The value only exists to probe that the pool
+            // is alive after migration.
+            let one: i32 = sqlx::query_scalar("SELECT 1")
                 .fetch_one(&pool)
                 .await
                 .expect("test postgres pool should answer after migration");
